fix: resolve all seven CodeQL code scanning alerts (#417)
* fix: resolve all seven CodeQL code scanning alerts

The first CodeQL run on main (#414) opened seven alerts: two go/zipslip,
four go/incorrect-integer-conversion and one
go/unhandled-writable-file-close. One is a real, deterministic bug; the
rest range from portability hazard to provability. All seven are closed
by construction rather than dismissed.

humanBytes wrapped uint64 byte counts to negative
-------------------------------------------------
cmd/skrog's humanBytes took int64, and all 19 call sites over the
engine's uint64 sizes bridged that with humanBytes(int64(x)). Measured:

    18446744073709551615 -> "-1 B"
     9223372036854775808 -> "-9223372036854775808 B"

A negative byte count, from a field that cannot be negative. The values
come out of the engine's own JSON, so the bound was never ours to
assume. humanBytes is now generic over int64|uint64, which removes the
conversion instead of range-checking it, and the 19 wrappers are gone.

formatValue performed the conversion it was guarding
----------------------------------------------------
`f == float64(int64(f))` decides whether int64(f) is safe by evaluating
int64(f), which Go leaves *undefined* out of range.

Measured on amd64: the overflow saturates to MinInt64, the round-trip
comparison fails, and the output is correct. The old code was not wrong
-- it was lucky, on one ISA. arm64 saturates to MaxInt64 instead, and
float64(MaxInt64) rounds back up to 2^63, so an input of exactly 2^63
round-trips successfully there and prints 9223372036854775807: a wrong
number, silently, for a value handed over intact. arm64 has been a
shipped target since #389. Now bounds-checked before converting, with
NaN and the infinities falling through to %g, which spells them the way
the exposition format does.

zipslip: the archive's string no longer reaches filepath.Join
--------------------------------------------------------------
Neither site could actually escape -- filepath.Base/path.Base plus a
fixed allowlist saw to that, and the zip side has had a traversal test
since #248. But the name still flowed from the archive into
filepath.Join, so nothing in the code said so: not to a reader, and not
to CodeQL. Both now resolve the match to the constant from
Binaries/EngineBinaries, so the path written is provably ours and the
untrusted string stops at a comparison.

The tar side had no traversal test at all, which is the more plausible
carrier -- a rootfs tarball comes from a registry, not from our release
page. It has one now, covering both the entry that path.Clean rejects
and the one it legitimately resolves into the engine's bin directory.

rotate.go dropped Close on a writable handle
---------------------------------------------
Joined into the returned error instead. Nothing has been written to the
handle at that point, so it should never have anything to report --
which is the argument for surfacing it if it ever does.

Verification
------------
The humanBytes test is a true negative control: it fails against the
shipped code, confirmed by running the old implementation directly. The
formatValue arm64 case is checked by the arm64 CI job, not by this
machine, and the comment says so. The zipslip tests are
characterization, not regression catches -- the behaviour did not
change there, only what can be proven about it.

go test ./..., lint.ps1, govulncheck and a GOOS=linux build of every
touched package are green. (internal/doctor does not build for linux;
that is pre-existing and unrelated, confirmed by stashing.)


* fix: stop narrowing to int64 in formatValue at all

Bounds-checking the conversion closed six of the seven alerts and opened
a new one at the same line. CodeQL was right to be unimpressed: the
value arrives as a uint64 from a strconv.ParseUint of /proc inside the
distro, and a guard written in the float domain is not something the
analysis can tie back to the integer it came from. The fix still
narrowed; it just narrowed carefully.

There is no need to narrow. math.Trunc answers "is this whole?" without
leaving float64, and strconv.FormatFloat prints the integer without an
int64 ever existing.

That also moves the boundary to the right place. It was 2^63, where
int64 stops; it is now 2^53, where float64 stops representing
consecutive integers. Past 2^53 the trailing digits of the plain form
would be invented, so the exponent form is the more honest rendering
rather than merely the safe one. 2^53 bytes is 9 PiB -- every value
this tool actually measures is far below it and still prints plain.


---------

+++ cmd/skrog/cache.go
@@ -193,7 +193,7 @@ func cacheStatus(ctx context.Context, r prune.DockerRunner, opts provision.Optio
 		return exitOK
 	}
 	fmt.Printf("cache      %s\nupstream   %s\nmirror     %s\ndata       %s\n",
-		runningWord(st.Running), st.Upstream, st.MirrorURL, humanBytes(int64(st.DataBytes)))
+		runningWord(st.Running), st.Upstream, st.MirrorURL, humanBytes(st.DataBytes))
 	// The one state worth calling out: a cache that is up and that nothing is
 	// pointed at looks healthy and does nothing.
 	if !st.Wired {
